WIP: Add a remote font configuration polling thread
The coolwsd.xml file can contain a URI of a JSON file on some server
that contains URIs of fonts. These fonts are downloaded. Just like the
remote configuration thing, the URIs are checked once a minute and the
JSON or the fonts are re-downloaded if their contents has changed.

If a font has been removed from the JSON file then the corresponding
downloaded could be removed, too. But there is no way to remove it
from core without restarting the whole COOL server, so we don't
bother.

We download the fonts into a fresh subdirectory of /tmp and then
bind-mound that read-only in the Kit processes using the same
(jail-relative) pathname, so that the code in vcl can open and mmap
such a font (see FreetypeFontFile::Map() in
vcl/unx/generic/glyphs/freetype_glyphcache.cxx) using the same
pathname.

Signed-off-by: Tor Lillqvist <tml@collabora.com>
Change-Id: If78058ddff5ed05c7a82d7ea465a7a414fd0d861